spring boot first commit

+++ journalapp/src/main/java/com/edigest/journalapp/services/journalentryservice.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 //import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.SQLOutput;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -28,7 +29,7 @@ public class journalentryservice {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = JournalEntryRepositery.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userservice.saveEntry(user);
+            userservice.saveuser(user);
         }
         catch(Exception e){
             System.out.println(e);
@@ -48,14 +49,20 @@ public class journalentryservice {
         return JournalEntryRepositery.findById(String.valueOf(id));
     }
 
-    public void deleteByid(ObjectId id, String userName){
-        User user = userservice.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userservice.saveEntry(user);
-        JournalEntryRepositery.deleteById(String.valueOf(id));
+    public boolean deleteByid(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userservice.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userservice.saveuser(user);
+                JournalEntryRepositery.deleteById(String.valueOf(id));
+            }
+
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting",e);
+        }
+        return removed;
     }
-//
-//    public List<JournalEntry> findByUsername(String userName){
-//
-//    }
 }
